# ratelimiter/docs/swagger.yaml
basePath: /
definitions:
  schemas.RateLimitRequest:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
    type: object
  schemas.RateLimitResponse:
    properties:
      is_allowed:
        type: boolean
      retry_after_seconds:
        type: integer
      tokens_capacity:
        type: integer
      tokens_remaining:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for rate limiting service
  title: Rate Limiter API
  version: "1.0"
paths:
  /api/v1/ratelimit/{entity_type}/{entity_id}/status:
    get:
      consumes:
      - application/json
      description: Returns the current rate limit status for the specified entity.
      parameters:
      - description: Entity type
        in: path
        name: entity_type
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RateLimitResponse'
      summary: Show rate limit status
      tags:
      - ratelimit
  /api/v1/ratelimit/consume:
    post:
      consumes:
      - application/json
      description: Consumes a rate limit for the specified entity.
      parameters:
      - description: Rate limit consume request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/schemas.RateLimitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.RateLimitResponse'
      summary: consume rate limit
      tags:
      - ratelimit
  /health:
    get:
      description: check server status
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Health check
      tags:
      - health
swagger: "2.0"